centralized proptypes, added comments
