+++ src/terra-dev-site/IssueForm/PackageSelect.jsx
@@ -1,12 +1,7 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import Field from 'terra-form-field';
 import Select from 'terra-form-select';
-
-const propTypes = {
-  packageList: PropTypes.arrayOf(PropTypes.string),
-  setPackage: PropTypes.func,
-};
+import propTypes from './Helper';
 
 const PackageSelect = (props) => {
   const {
